TG-96: moved Layout component to a new folder (/routes), ceated an employees.js to add the routes related to this user type. employees.js imported with lazy in the layout component, Suspense applied.
