+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import Layout from 'Components/Layout';
+import Layout from 'routes';
 import store from 'redux/store';
 import './index.css';
